@@ -18,11 +18,11 @@
         "optimize-autoloader": true
     },
     "require": {
-        "php": ">=7.2"
+        "php": ">=7.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5",
-        "blackfire/php-sdk": "^1.18"
+        "phpunit/phpunit": "^8.5",
+        "blackfire/php-sdk": "^1.21"
     },
     "autoload": {
         "psr-4": {
