@@ -6,16 +6,9 @@ namespace PTS\Events;
 abstract class BaseEvents
 {
     /** @var array[][] */
-    protected $listeners = [];
-    /** @var ResolveHandler */
-    protected $handler;
-    /** @var bool */
-    protected $sorted = false;
-
-    public function __construct(ResolveHandler $handler = null)
-    {
-        $this->handler = $handler ?? new ResolveHandler;
-    }
+    protected array $listeners = [];
+    /** @var bool[] */
+    protected array $sorted = [];
 
     public function getListeners(): array
     {
@@ -31,17 +24,9 @@ abstract class BaseEvents
         }
     }
 
-    /**
-     * @param string $name
-     * @param callable $handler
-     * @param int $priority
-     * @param array $extraArguments
-     *
-     * @return $this
-     */
-    public function on(string $name, callable $handler, int $priority = 50, array $extraArguments = [])
+    public function on(string $name, callable $handler, int $priority = 50, array $extraArguments = []): self
     {
-        $handlerId = $this->handler->getKey($handler);
+        $handlerId = $this->getHandlerId($handler);
         $this->listeners[$name][$priority][$handlerId] = [
             'handler' => $handler,
             'extraArguments' => $extraArguments,
@@ -53,28 +38,15 @@ abstract class BaseEvents
         return $this;
     }
 
-    /**
-     * @param string $name
-     * @param callable $handler
-     * @param int $priority
-     * @param array $extraArguments
-     *
-     * @return $this
-     */
     public function once(string $name, callable $handler, int $priority = 50, array $extraArguments = []): self
     {
         $this->on($name, $handler, $priority, $extraArguments);
-        $handlerId = $this->handler->getKey($handler);
+        $handlerId = $this->getHandlerId($handler);
         $this->listeners[$name][$priority][$handlerId]['once'] = true;
 
         return $this;
     }
 
-    /**
-     * @param string $eventName
-     *
-     * @return $this
-     */
     protected function offAll(string $eventName): self
     {
         $hasEvent = $this->listeners[$eventName] ?? false;
@@ -85,13 +57,6 @@ abstract class BaseEvents
         return $this;
     }
 
-    /**
-     * @param string $eventName
-     * @param callable|null $handler
-     * @param int|null $priority
-     *
-     * @return $this
-     */
     public function off(string $eventName, callable $handler = null, int $priority = null): self
     {
         if ($handler === null) {
@@ -109,16 +74,9 @@ abstract class BaseEvents
         return $this;
     }
 
-    /**
-     * @param string $eventName
-     * @param callable $handler
-     * @param int $priority
-     *
-     * @return $this
-     */
     protected function offHandlerWithPriority(string $eventName, callable $handler, int $priority = 50): self
     {
-        $handlerId = $this->handler->getKey($handler);
+        $handlerId = $this->getHandlerId($handler);
 
         if (isset($this->listeners[$eventName][$priority][$handlerId])) {
             unset($this->listeners[$eventName][$priority][$handlerId]);
@@ -128,15 +86,9 @@ abstract class BaseEvents
         return $this;
     }
 
-    /**
-     * @param string $eventName
-     * @param callable $handler
-     *
-     * @return $this
-     */
     protected function offHandlerWithoutPriority(string $eventName, callable $handler): self
     {
-        $handlerId = $this->handler->getKey($handler);
+        $handlerId = $this->getHandlerId($handler);
 
         foreach ($this->listeners[$eventName] as $currentPriority => $handlers) {
             foreach ($handlers as $currentHandlerId => $paramsHandler) {
@@ -161,34 +113,6 @@ abstract class BaseEvents
         }
     }
 
-    /**
-     * @param string $name
-     * @param array $arguments
-     * @param mixed $value
-     *
-     * @return mixed
-     */
-    protected function trigger(string $name, array $arguments = [], $value = null)
-    {
-        $event = $this->listeners[$name] ?? false;
-        if ($event) {
-            $this->sortListeners($name);
-            foreach ($this->listeners[$name] as $handlers) {
-                foreach ($handlers as $paramsHandler) {
-                    $callArgs = $this->getCallArgs($arguments, $paramsHandler['extraArguments'], $value);
-                    $value = $paramsHandler['handler'](...$callArgs);
-                    $this->offOnce($paramsHandler, $name);
-                }
-            }
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param array $paramsHandler
-     * @param string $name
-     */
     protected function offOnce(array $paramsHandler, string $name): void
     {
         if ($paramsHandler['once'] ?? false) {
@@ -196,19 +120,18 @@ abstract class BaseEvents
         }
     }
 
-    /**
-     * @param array $arguments
-     * @param array $extraArguments
-     * @param mixed $value
-     *
-     * @return array
-     */
-    protected function getCallArgs(array $arguments, array $extraArguments, $value = null): array
+    public function getHandlerId(callable $handler): string
     {
-        foreach ($extraArguments as $name => $val) {
-            $arguments[] = $val;
+        if (is_array($handler)) {
+            [$className, $method] = $handler;
+
+            if (is_object($className)) {
+                $className = get_class($className);
+            }
+
+            return "{$className}::{$method}";
         }
 
-        return $arguments;
+        return is_string($handler) ? $handler : spl_object_hash($handler);
     }
 }
