@@ -5,16 +5,22 @@ namespace PTS\Events;
 
 class Events extends BaseEvents implements EventsInterface
 {
-    /**
-     * @param string $name
-     * @param array $arguments
-     *
-     * @return $this
-     */
-    public function emit(string $name, array $arguments = [])
+
+    public function emit(string $name, array $arguments = []): self
     {
+        $event = $this->listeners[$name] ?? false;
+        if (!$event) {
+            return $this;
+        }
+
         try {
-            $this->trigger($name, $arguments);
+            $this->sortListeners($name);
+            foreach ($this->listeners[$name] as $handlers) {
+                foreach ($handlers as $paramsHandler) {
+                    $paramsHandler['handler'](...$arguments, ...$paramsHandler['extraArguments']);
+                    $this->offOnce($paramsHandler, $name);
+                }
+            }
         } catch (StopPropagation $e) {
             return $this;
         }
