@@ -5,27 +5,26 @@ namespace PTS\Events;
 
 class Filters extends BaseEvents implements FiltersInterface
 {
-    /**
-     * @param string $name
-     * @param mixed $value
-     * @param array $arguments
-     *
-     * @return mixed
-     */
+
     public function filter(string $name, $value, array $arguments = [])
     {
+        $filter = $this->listeners[$name] ?? false;
+        if ($filter === false) {
+            return $value;
+        }
+
         try {
-            return $this->trigger($name, $arguments, $value);
+            $this->sortListeners($name);
+            foreach ($this->listeners[$name] as $handlers) {
+                foreach ($handlers as $paramsHandler) {
+                    $value = $paramsHandler['handler']($value, ...$arguments, ...$paramsHandler['extraArguments']);
+                    $this->offOnce($paramsHandler, $name);
+                }
+            }
         } catch (StopPropagation $e) {
             return $e->getValue();
         }
-    }
 
-    protected function getCallArgs(array $arguments, array $extraArguments, $value = null): array
-    {
-        $arguments = parent::getCallArgs($arguments, $extraArguments, $value);
-        array_unshift($arguments, $value);
-
-        return $arguments;
+        return $value;
     }
 }
